backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers